* origin/master: (162 commits)
  BF: monkey patch scipy with broken fit for 2 fixed params (e.g. floc and fscale)
  BF: match_distributino - use "resolved" dist_gen, which might original be a tuple
  BF: workaround for sphinx -- import Process as private __Process from psutil
  BF: fix legend in plot match distributions
  ENH: suppress Python warnings while match_distribution
  ENH/BF: compute unique only if not known + handle all obscure cases via ndarray(list(set()))
  ENH: just an additional externals check to see if np.unique is flawed
  BF: h5py external is required for results_backend="hdf5"
  ENH: run unittest-py3 for origin/master and tiziano/master +store output for examples +safe guard WTF call
  PY3: make SequenceCollectable.unique faster.
  BF: assure different filenames in the SL's _proc_block by providing index of the block (should fix gh 87)
  ENH: make h5py version available
  DOC: Start a changelog for 2.2?
  BF: Remove special case for builtin reconstructors in HDF5 (Closes gh-86)
  Document the HDF5 cannot handle slice()
  DOC: Add reference to Emanuele's conference paper.
  BF: hdf5 -- flatten and save attr 'shape' for object ndarrays (Closes gh-84)
  ENH: fix up the searchlight test for Swaroop's intriguing measure
  Modified searchlight unittest to show hdf5 & native modes differ
  ENH: results_backend to alter how results are passed from pprocess
  ...

Conflicts: primarily due to added in master .ca.roi_feature_ids and heavy
           RF of the test_spatial_searchlight  in _tent/mnn branch
	mvpa2/measures/gnbsearchlight.py
	mvpa2/tests/test_searchlight.py

+++ mvpa2/mappers/som.py
@@ -107,7 +107,6 @@ class SimpleSOMMapper(Mapper):
             for s in samples:
                 # determine closest unit (as element coordinate)
                 b = self._get_bmu(s)
-
                 # train all units at once by unfolding the kernel (from the
                 # single quadrant that is precomputed), cutting it to the
                 # right shape and simply multiply it to the difference of target
@@ -184,9 +183,8 @@ class SimpleSOMMapper(Mapper):
         """
         # TODO expose distance function as parameter
         loc = np.argmin(((self.K - sample) ** 2).sum(axis=2))
-
         # assumes 2D Kohonen layer
-        return (np.divide(loc, self.kshape[1]), loc % self.kshape[1])
+        return (np.divide(loc, self.kshape[1]).astype('int'), loc % self.kshape[1])
 
 
     def _forward_data(self, data):
